+++ native/src/main/java/io/ballerina/stdlib/http/api/HttpService.java
@@ -19,6 +19,7 @@ package io.ballerina.stdlib.http.api;
 
 import io.ballerina.runtime.api.flags.SymbolFlags;
 import io.ballerina.runtime.api.types.MethodType;
+import io.ballerina.runtime.api.types.ObjectType;
 import io.ballerina.runtime.api.types.ServiceType;
 import io.ballerina.runtime.api.utils.StringUtils;
 import io.ballerina.runtime.api.values.BMap;
@@ -33,9 +34,13 @@ import io.ballerina.stdlib.http.uri.parser.Literal;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.File;
 import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -55,10 +60,11 @@ public class HttpService {
     private static final BString CORS_FIELD = StringUtils.fromString("cors");
     private static final BString VERSIONING_FIELD = StringUtils.fromString("versioning");
     private static final BString HOST_FIELD = StringUtils.fromString("host");
+    private static final BString MEDIA_TYPE_SUBTYPE_PREFIX = StringUtils.fromString("mediaTypeSubtypePrefix");
+    private static final BString TREAT_NILABLE_AS_OPTIONAL = StringUtils.fromString("treatNilableAsOptional");
 
     private BObject balService;
     private List<HttpResource> resources;
-    private List<HttpResource> upgradeToWebSocketResources;
     private List<String> allAllowedMethods;
     private String basePath;
     private CorsHeaders corsHeaders;
@@ -67,6 +73,9 @@ public class HttpService {
     private BMap<BString, Object> compression;
     private String hostName;
     private String chunkingConfig;
+    private String mediaTypeSubtypePrefix;
+    private String introspectionResourcePath;
+    private boolean treatNilableAsOptional = true;
 
     protected HttpService(BObject service, String basePath) {
         this.balService = service;
@@ -138,6 +147,22 @@ public class HttpService {
         return hostName;
     }
 
+    public void setMediaTypeSubtypePrefix(String mediaTypeSubtypePrefix) {
+        this.mediaTypeSubtypePrefix = mediaTypeSubtypePrefix;
+    }
+
+    public String getMediaTypeSubtypePrefix() {
+        return mediaTypeSubtypePrefix;
+    }
+
+    public void setTreatNilableAsOptional(boolean treatNilableAsOptional) {
+        this.treatNilableAsOptional = treatNilableAsOptional;
+    }
+
+    public boolean isTreatNilableAsOptional() {
+        return treatNilableAsOptional;
+    }
+
     public String getBasePath() {
         return basePath;
     }
@@ -185,6 +210,11 @@ public class HttpService {
             httpService.setChunkingConfig(serviceConfig.get(HttpConstants.ANN_CONFIG_ATTR_CHUNKING).toString());
             httpService.setCorsHeaders(CorsHeaders.buildCorsHeaders(serviceConfig.getMapValue(CORS_FIELD)));
             httpService.setHostName(serviceConfig.getStringValue(HOST_FIELD).getValue().trim());
+            if (serviceConfig.containsKey(MEDIA_TYPE_SUBTYPE_PREFIX)) {
+                httpService.setMediaTypeSubtypePrefix(serviceConfig.getStringValue(MEDIA_TYPE_SUBTYPE_PREFIX)
+                        .getValue().trim());
+            }
+            httpService.setTreatNilableAsOptional(serviceConfig.getBooleanValue(TREAT_NILABLE_AS_OPTIONAL));
         } else {
             httpService.setHostName(HttpConstants.DEFAULT_HOST);
         }
@@ -195,21 +225,34 @@ public class HttpService {
 
     private static void processResources(HttpService httpService) {
         List<HttpResource> httpResources = new ArrayList<>();
-        for (MethodType resource :
-                ((ServiceType) httpService.getBalService().getType()).getResourceMethods()) {
+        for (MethodType resource : ((ServiceType) httpService.getBalService().getType()).getResourceMethods()) {
             if (!SymbolFlags.isFlagOn(resource.getFlags(), SymbolFlags.RESOURCE)) {
                 continue;
             }
-            HttpResource httpResource = HttpResource.buildHttpResource(resource, httpService);
-            try {
-                httpService.getUriTemplate().parse(httpResource.getPath(), httpResource,
-                                                   new HttpResourceElementFactory());
-            } catch (URITemplateException | UnsupportedEncodingException e) {
-                throw new BallerinaConnectorException(e.getMessage());
-            }
-            httpResources.add(httpResource);
+            updateResourceTree(httpService, httpResources, HttpResource.buildHttpResource(resource, httpService));
         }
+
+        httpService.getIntrospectionDocName().ifPresent(openApiDocName -> {
+            String filePath = "resources/ballerina/http/" + openApiDocName + ".json";
+            File tempFile = new File(filePath);
+            if (tempFile.exists()) {
+                updateResourceTree(httpService, httpResources, new HttpIntrospectionResource(httpService, filePath));
+            } else {
+                log.debug("OpenAPI specification document does not exist in path: '" + filePath + "'");
+            }
+        });
         httpService.setResources(httpResources);
+    }
+
+    private static void updateResourceTree(HttpService httpService, List<HttpResource> httpResources,
+                                           HttpResource httpResource) {
+        try {
+            httpService.getUriTemplate().parse(httpResource.getPath(), httpResource, new HttpResourceElementFactory());
+        } catch (URITemplateException | UnsupportedEncodingException e) {
+            throw new BallerinaConnectorException(e.getMessage());
+        }
+        httpResource.setTreatNilableAsOptional(httpService.isTreatNilableAsOptional());
+        httpResources.add(httpResource);
     }
 
     private static BMap getHttpServiceConfigAnnotation(BObject service) {
@@ -221,5 +264,26 @@ public class HttpService {
                                                      String annotationName) {
         String key = packagePath.replaceAll(HttpConstants.REGEX, HttpConstants.SINGLE_SLASH);
         return (BMap) (service.getType()).getAnnotation(StringUtils.fromString(key + ":" + annotationName));
+    }
+
+    protected String getIntrospectionResourcePathHeaderValue() {
+        return this.introspectionResourcePath;
+    }
+
+    protected void setIntrospectionResourcePathHeaderValue(String introspectionResourcePath) {
+        this.introspectionResourcePath = introspectionResourcePath;
+    }
+
+    protected Optional<String> getIntrospectionDocName() {
+        ObjectType objType = this.balService.getType();
+        if (Objects.nonNull(objType)) {
+            return objType.getAnnotations().entrySet().stream()
+                    .filter(e -> e.getKey().toString().contains(HttpConstants.ANN_NAME_HTTP_INTROSPECTION_DOC_CONFIG))
+                    .findFirst()
+                    .map(Map.Entry::getValue)
+                    .filter(e -> e instanceof BMap)
+                    .map(e -> ((BMap) e).getStringValue(HttpConstants.ANN_FIELD_DOC_NAME).getValue().trim());
+        }
+        return Optional.empty();
     }
 }