+++ native/src/main/java/io/ballerina/stdlib/http/api/service/signature/ParamHandler.java
@@ -42,7 +42,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import static io.ballerina.stdlib.http.api.HttpConstants.ANN_NAME_CACHE_CONFIG;
+import static io.ballerina.stdlib.http.api.HttpConstants.ANN_NAME_CACHE;
 import static io.ballerina.stdlib.http.api.HttpConstants.ANN_NAME_CALLER_INFO;
 import static io.ballerina.stdlib.http.api.HttpConstants.ANN_NAME_HEADER;
 import static io.ballerina.stdlib.http.api.HttpConstants.ANN_NAME_PAYLOAD;
@@ -78,8 +78,8 @@ public class ParamHandler {
             ModuleUtils.getHttpPackageIdentifier() + COLON + ANN_NAME_CALLER_INFO;
     public static final String PAYLOAD_ANNOTATION = ModuleUtils.getHttpPackageIdentifier() + COLON + ANN_NAME_PAYLOAD;
     public static final String HEADER_ANNOTATION = ModuleUtils.getHttpPackageIdentifier() + COLON + ANN_NAME_HEADER;
-    public static final String CACHE_CONFIG_ANNOTATION = ModuleUtils.getHttpPackageIdentifier() + COLON
-            + ANN_NAME_CACHE_CONFIG;
+    public static final String CACHE_ANNOTATION = ModuleUtils.getHttpPackageIdentifier() + COLON
+            + ANN_NAME_CACHE;
 
     public ParamHandler(ResourceMethodType resource, int pathParamCount) {
         this.resource = resource;
@@ -216,7 +216,7 @@ public class ParamHandler {
             headerParam.setHeaderName(value);
         } else {
             // if the name field is not stated, use the param token as header key
-            headerParam.setHeaderName(paramName);
+            headerParam.setHeaderName(HttpUtil.unescapeAndEncodeValue(paramName));
         }
         this.headerParams.add(headerParam);
     }
@@ -234,12 +234,14 @@ public class ParamHandler {
                 if (type.getTag() == TypeTags.NULL_TAG) {
                     continue;
                 }
-                QueryParam queryParam = new QueryParam(type, balResource.getParamNames()[index], index, true);
+                QueryParam queryParam = new QueryParam(type, HttpUtil.unescapeAndEncodeValue(
+                        balResource.getParamNames()[index]), index, true);
                 this.queryParams.add(queryParam);
                 break;
             }
         } else {
-            QueryParam queryParam = new QueryParam(parameterType, balResource.getParamNames()[index], index, false);
+            QueryParam queryParam = new QueryParam(parameterType, HttpUtil.unescapeAndEncodeValue(
+                    balResource.getParamNames()[index]), index, false);
             this.queryParams.add(queryParam);
         }
     }